Remove unnecessary uint32_t casts for draw and drawIndexed calls.

+++ src/gpu/ops/AtlasTextOp.cpp
@@ -90,8 +90,7 @@ void AtlasTextOp::onDraw(RenderPass* renderPass) {
     } else {
       numIndicesPerQuad = RectDrawOp::IndicesPerNonAAQuad;
     }
-    renderPass->drawIndexed(PrimitiveType::Triangles,
-                            static_cast<uint32_t>(rectCount * numIndicesPerQuad));
+    renderPass->drawIndexed(PrimitiveType::Triangles, rectCount * numIndicesPerQuad);
   } else {
     renderPass->draw(PrimitiveType::TriangleStrip, 4);
   }
